recipes-trustx/userdata/pki-native: Test for existing key
When rebuilding the pki-recipe this step may fail if the symlink already
exists. Testing for its existence beforehand avoids the recipe failing.

+++ recipes-trustx/userdata/pki-native.bbappend
@@ -1,4 +1,6 @@
 
 do_compile:append(){
-    ln -s ${TEST_CERT_DIR}/ssig_subca.cert ${TEST_CERT_DIR}/ssig_subca.crt
+	if [ ! -e "${TEST_CERT_DIR}/ssig_subca.crt" ]; then
+		ln -s ${TEST_CERT_DIR}/ssig_subca.cert ${TEST_CERT_DIR}/ssig_subca.crt
+	fi
 }
